Refactor: post 관련 컴포넌트 구조 수정
