+++ src/components/feature/board/navigation-button/NavigationButton.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import { ArrowRightIcon } from '@radix-ui/react-icons'
 import { usePathname, useRouter } from 'next/navigation'
