+++ src/components/feature/board/navigation-button/index.ts
@@ -1,0 +1,1 @@
+export { BoardNavigationButton } from './NavigationButton'
