+++ src/components/feature/index.ts
@@ -2,5 +2,4 @@ export * from './activity'
 export * from './admin'
 export * from './post'
 export * from './skeleton'
-
-export * from './image-input'
+export * from './board'
