+++ src/components/feature/post/content/index.ts
@@ -1,0 +1,1 @@
+export { default as PostContent } from './Content'
