--- src/components/feature/post/create-post-form/image-input/ImageInput.tsx
@@ -1,20 +1,0 @@
-import { ControllerRenderProps } from 'react-hook-form'
-
-import { Input } from '@/components/ui/input'
-
-type ImageInputProps = {
-  field: ControllerRenderProps
-}
-
-export const ImageInput = ({ field }: ImageInputProps) => {
-  return (
-    <Input
-      accept=".jpg, .jpeg"
-      type="file"
-      multiple={false}
-      onChange={(e) =>
-        field.onChange(e.target.files ? e.target.files[0] : null)
-      }
-    />
-  )
-}
