+++ src/components/feature/post/index.ts
@@ -1,3 +1,5 @@
 export * from './create-post-form'
 export * from './form-field'
 export * from './table'
+export * from './content'
+export * from './image-input'
